Move basic preferences controller to preferences module
