--- app/views/preferences/_update_success_flash.html.erb
@@ -1,1 +1,0 @@
-<%= t ".message" %>
