+++ app/views/preferences/basic_preferences/_update_success_flash.html.erb
@@ -1,0 +1,5 @@
+<%= if I18n.exists? "preferences.update_success_flash.message"
+      t "preferences.update_success_flash.message"
+    else
+      t ".message"
+    end %>
