+++ app/views/preferences/basic_preferences/show.html.erb
