+++ config/locales/en.yml
@@ -1949,23 +1949,24 @@ en:
       flash changed: "Your password has been changed."
       flash token bad: "Did not find that token, check the URL maybe?"
   preferences:
-    show:
-      title: My Preferences
-      preferred_site_color_scheme: Preferred Website Color Scheme
-      site_color_schemes:
-        auto: Auto
-        light: Light
-        dark: Dark
-      preferred_map_color_scheme: Preferred Map Color Scheme
-      map_color_schemes:
-        auto: Auto
-        light: Light
-        dark: Dark
-      save: Update Preferences
-    update:
-      failure: Couldn't update preferences.
-    update_success_flash:
-      message: Preferences updated.
+    basic_preferences:
+      show:
+        title: My Preferences
+        preferred_site_color_scheme: Preferred Website Color Scheme
+        site_color_schemes:
+          auto: Auto
+          light: Light
+          dark: Dark
+        preferred_map_color_scheme: Preferred Map Color Scheme
+        map_color_schemes:
+          auto: Auto
+          light: Light
+          dark: Dark
+        save: Update Preferences
+      update:
+        failure: Couldn't update preferences.
+      update_success_flash:
+        message: Preferences updated.
   profiles:
     show:
       title: Edit Profile
